+++ tests/unit/test_variable.py
@@ -10,8 +10,13 @@ from gamspy import (
     Model,
     Options,
     Parameter,
+    Product,
+    Sand,
     Sense,
     Set,
+    Smax,
+    Smin,
+    Sor,
     Sum,
     Variable,
     VariableType,
@@ -542,3 +547,187 @@ def test_alternative_syntax():
     assert v2.getAssignment() == "v2.prior(i) = 5;"
     v[i].stage = 5
     assert v.getAssignment() == "v.stage(i) = 5;"
+
+
+def test_alternative_operation_syntax():
+    m = Container()
+
+    i = Set(m)
+    j = Set(m)
+    x = Variable(m, domain=[i, j])
+    y = Variable(m)
+
+    # Test sum
+    with pytest.raises(ValidationError):
+        y.sum()
+
+    expr = x.sum()
+    expr2 = Sum((i, j), x[i, j])
+    assert expr.gamsRepr() == expr2.gamsRepr()
+
+    expr = x.sum(i)
+    expr2 = Sum(i, x[i, j])
+    assert expr.gamsRepr() == expr2.gamsRepr()
+
+    expr = x.sum(i, j)
+    expr2 = Sum((i, j), x[i, j])
+    assert expr.gamsRepr() == expr2.gamsRepr()
+
+    # Test product
+    with pytest.raises(ValidationError):
+        y.product()
+
+    expr = x.product()
+    expr2 = Product((i, j), x[i, j])
+    assert expr.gamsRepr() == expr2.gamsRepr()
+
+    expr = x.product(i)
+    expr2 = Product(i, x[i, j])
+    assert expr.gamsRepr() == expr2.gamsRepr()
+
+    expr = x.product(i, j)
+    expr2 = Product((i, j), x[i, j])
+    assert expr.gamsRepr() == expr2.gamsRepr()
+
+    # Test smin
+    with pytest.raises(ValidationError):
+        y.smin()
+
+    expr = x.smin()
+    expr2 = Smin((i, j), x[i, j])
+    assert expr.gamsRepr() == expr2.gamsRepr()
+
+    expr = x.smin(i)
+    expr2 = Smin(i, x[i, j])
+    assert expr.gamsRepr() == expr2.gamsRepr()
+
+    expr = x.smin(i, j)
+    expr2 = Smin((i, j), x[i, j])
+    assert expr.gamsRepr() == expr2.gamsRepr()
+
+    # Test smax
+    with pytest.raises(ValidationError):
+        y.smax()
+
+    expr = x.smax()
+    expr2 = Smax((i, j), x[i, j])
+    assert expr.gamsRepr() == expr2.gamsRepr()
+
+    expr = x.smax(i)
+    expr2 = Smax(i, x[i, j])
+    assert expr.gamsRepr() == expr2.gamsRepr()
+
+    expr = x.smax(i, j)
+    expr2 = Smax((i, j), x[i, j])
+    assert expr.gamsRepr() == expr2.gamsRepr()
+
+    # Test sand
+    with pytest.raises(ValidationError):
+        y.sand()
+
+    expr = x.sand()
+    expr2 = Sand((i, j), x[i, j])
+    assert expr.gamsRepr() == expr2.gamsRepr()
+
+    expr = x.sand(i)
+    expr2 = Sand(i, x[i, j])
+    assert expr.gamsRepr() == expr2.gamsRepr()
+
+    expr = x.sand(i, j)
+    expr2 = Sand((i, j), x[i, j])
+    assert expr.gamsRepr() == expr2.gamsRepr()
+
+    # Test sor
+    with pytest.raises(ValidationError):
+        y.sor()
+
+    expr = x.sor()
+    expr2 = Sor((i, j), x[i, j])
+    assert expr.gamsRepr() == expr2.gamsRepr()
+
+    expr = x.sor(i)
+    expr2 = Sor(i, x[i, j])
+    assert expr.gamsRepr() == expr2.gamsRepr()
+
+    expr = x.sor(i, j)
+    expr2 = Sor((i, j), x[i, j])
+    assert expr.gamsRepr() == expr2.gamsRepr()
+
+    ### ImplicitParameter
+    # Test sum
+    expr = x[i, j].sum()
+    expr2 = Sum((i, j), x[i, j])
+    assert expr.gamsRepr() == expr2.gamsRepr()
+
+    expr = x[i, j].sum(i)
+    expr2 = Sum(i, x[i, j])
+    assert expr.gamsRepr() == expr2.gamsRepr()
+
+    expr = x[i, j].sum(i, j)
+    expr2 = Sum((i, j), x[i, j])
+    assert expr.gamsRepr() == expr2.gamsRepr()
+
+    # Test product
+    expr = x[i, j].product()
+    expr2 = Product((i, j), x[i, j])
+    assert expr.gamsRepr() == expr2.gamsRepr()
+
+    expr = x[i, j].product(i)
+    expr2 = Product(i, x[i, j])
+    assert expr.gamsRepr() == expr2.gamsRepr()
+
+    expr = x[i, j].product(i, j)
+    expr2 = Product((i, j), x[i, j])
+    assert expr.gamsRepr() == expr2.gamsRepr()
+
+    # Test smin
+    expr = x[i, j].smin()
+    expr2 = Smin((i, j), x[i, j])
+    assert expr.gamsRepr() == expr2.gamsRepr()
+
+    expr = x[i, j].smin(i)
+    expr2 = Smin(i, x[i, j])
+    assert expr.gamsRepr() == expr2.gamsRepr()
+
+    expr = x[i, j].smin(i, j)
+    expr2 = Smin((i, j), x[i, j])
+    assert expr.gamsRepr() == expr2.gamsRepr()
+
+    # Test smax
+    expr = x[i, j].smax()
+    expr2 = Smax((i, j), x[i, j])
+    assert expr.gamsRepr() == expr2.gamsRepr()
+
+    expr = x[i, j].smax(i)
+    expr2 = Smax(i, x[i, j])
+    assert expr.gamsRepr() == expr2.gamsRepr()
+
+    expr = x[i, j].smax(i, j)
+    expr2 = Smax((i, j), x[i, j])
+    assert expr.gamsRepr() == expr2.gamsRepr()
+
+    # Test sand
+    expr = x[i, j].sand()
+    expr2 = Sand((i, j), x[i, j])
+    assert expr.gamsRepr() == expr2.gamsRepr()
+
+    expr = x[i, j].sand(i)
+    expr2 = Sand(i, x[i, j])
+    assert expr.gamsRepr() == expr2.gamsRepr()
+
+    expr = x[i, j].sand(i, j)
+    expr2 = Sand((i, j), x[i, j])
+    assert expr.gamsRepr() == expr2.gamsRepr()
+
+    # Test sor
+    expr = x[i, j].sor()
+    expr2 = Sor((i, j), x[i, j])
+    assert expr.gamsRepr() == expr2.gamsRepr()
+
+    expr = x[i, j].sor(i)
+    expr2 = Sor(i, x[i, j])
+    assert expr.gamsRepr() == expr2.gamsRepr()
+
+    expr = x[i, j].sor(i, j)
+    expr2 = Sor((i, j), x[i, j])
+    assert expr.gamsRepr() == expr2.gamsRepr()